Fix OpenSSLContext to work correctly with OpenSSL 1.1.1
The previous code only worked with 1.1.0j or older.
Now the code works with 1.1.0j and OpenSSL 1.1.1.

Adjusted ClientServerTest to be more graceful in case
of errors, i.e. failing tests instead of crashing.

Test-Information:

Tested that without the changes, the tests pass with
OpenSSL 1.1.0j and test fail or crash with OpenSSL
1.1.1 and OpenSSL 1.1.1a.

Tested that with the changes, the tests pass with OpenSSL
1.1.0j, OpenSSL 1.1.1, and OpenSSL 1.1.1a.

Tested on macOS 10.14.2 with system clang.

Change-Id: Ic63774049727f6d949153166f63a8545e9a24892