Fix Python 3 Unicode issues related to Windows packaging
Test-Information:

Without these fixes, the build of a Windows MSI package
is not possible on Windows Server 2012. This is due to
the fact that on Windows you must force UTF8 coding
on file open as it defaults to Windows specific codec
for everything.

Tested by building a Windows MSI package on Windows
Server 2012.

Change-Id: I32664824188775f5ba27d9644fbbf33bf7094dfa