Add flag to signal final XML data
XML parse now takes an optional boolean that will signal that with the
data that is fed into the parser, the document should now be complete.
This will allow the parser to reject partial documents.

Test-Information:

Updated unit tests pass.
Checked that default behaviour wasn't changed.
Tested with various partial and complete documents.

Change-Id: Ide7c2e47c49d5667f1febcb23da366e96d0dbc21