Support application-supplied logging
This adds a method to set a logging callback.  If such a callback is
set, all SWIFT_LOG calls will invoke this callback instead of writing to
either stderr or the swift logging file.

Test-Information:

Updated unit tests pass.
Checked that logs generated by Swift and Sluift (which do not set the
callback) resulted in logging in the expected location.

Change-Id: I0eb2a1057aa77839e1b8d5f320205eb9d5fdc253