Remove std::endl from SWIFT_LOG calls
The std::endl is now added by ~Log, but only for output to stderr or a
log file.  Calls to the Android logging system or manually set callbacks
will not include the newline in the logging output.

JIRA: SWIFT-430

Test-Information:

Unit tests pass on Debian 9
Checked that running Swift with logging to stderr still had a newline.

Change-Id: I096fdba78a3b8f87db2097951c28c528592183e8